+++ test/user.clj
@@ -1,0 +1,4 @@
+(ns user
+  (:require [clj-commons.pretty.repl :as repl]))
+
+(repl/install-pretty-exceptions)