+++ test/my/example/ring_handler.clj
@@ -1,15 +1,20 @@
 (ns my.example.ring-handler
-  (:require [net.lewisship.trace :refer [trace]]))
+  (:require [clojure.string :as string]
+            [net.lewisship.trace :as trace :refer [trace]]))
+
+(trace/setup-default)
+
+(defn handle-resource-request [uri])
 
 (defn handle-request
   [request]
-  (trace
-    :method (:request-method request)
-    :uri (:uri request))
-  ;; Off to do some request handler type things
-  )
+  (if (string/ends-with? #trace/result (:uri request) "/")
+    {:status 401
+     :body   "Invalid request"})
+  (handle-resource-request (:uri request)))
 
 (comment
   (handle-request {:request-method :get :uri "/status"})
 
+  *data-readers*
   )