+++ CHANGES.md
@@ -3,6 +3,8 @@
 Added the ability to enable tracing in specific namespaces even when the global tracing flag
 is false.
 
+Migrated to org.clj-commons/pretty.
+
 # v1.0 - 10 Mar 2022
 
 Initial release.