+++ CHANGES.md
@@ -5,6 +5,9 @@ is false.
 
 Migrated to org.clj-commons/pretty.
 
+Added new `net.lewisship.bench` namespace with useful wrappers around 
+[criterium](https://github.com/hugoduncan/criterium) for benchmarking.
+
 # v1.0 - 10 Mar 2022
 
 Initial release.