+++ CHANGES.md
@@ -3,6 +3,11 @@
 Added the ability to enable tracing in specific namespaces even when the global tracing flag
 is false.
 
+Migrated to org.clj-commons/pretty.
+
+Added new `net.lewisship.bench` namespace with useful wrappers around 
+[criterium](https://github.com/hugoduncan/criterium) for benchmarking.
+
 # v1.0 - 10 Mar 2022
 
 Initial release.